% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_files_and_run_function.R
\name{watch_files_and_run_function}
\alias{watch_files_and_run_function}
\title{Watch files for changes and run a function}
\usage{
watch_files_and_run_function(fn, ..., delay_time = 1)
}
\arguments{
\item{fn}{a function to run when the files have changed}

\item{...}{the files that you want to monitor for changes, see details}

\item{delay_time}{how often (in seconds) to check for changes to files,
defaults to 1 second}
}
\description{
When any files that you pass in change, run a function once, then wait for
further changes. If you want to run code a separate process, see
\code{watch_files_and_start_task()}.
}
\details{
The files that you want to monitor can either be passed in as vectors
of file names, or functions that returns files. It must point to files
rather than entire directories, but the files don't necessarily have to
exist. The functions are re-evaluated for each loop, so if you want to
find all files in a directory you can use a function, see examples.
}
\examples{
\dontrun{
# watch all of the files in the R directory, and the DESCRIPTION file for
# changes, and print a message to the screen.
watch_files_and_run_function(
  \() print(files have changed),
  "DESCRIPTION",
  \() dir("R", "*.R")
)
}
}
