% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_files_and_start_task.R
\name{watch_files_and_start_task}
\alias{watch_files_and_start_task}
\title{Watch files for changes and run a function}
\usage{
watch_files_and_start_task(fn, ..., delay_time = 1)
}
\arguments{
\item{fn}{a function to run in a separate process when the files have changed}

\item{...}{the files that you want to monitor for changes, see details}

\item{delay_time}{how often (in seconds) to check for changes to files,
defaults to 1 second}
}
\description{
When any files that you pass in change, a new R process is started, and the
function that is passed in is started in that process. Previously running
tasks are killed, so there is only instance of one process running at any
given time.
}
\details{
The files that you want to monitor can either be passed in as vectors
of file names, or functions that returns files. It must point to files
rather than entire directories, but the files don't necessarily have to
exist. The functions are re-evaluated for each loop, so if you want to
find all files in a directory you can use a function, see examples.
}
\examples{
\dontrun{
# watch all of the files in the R directory, and the DESCRIPTION file for
# changes, and print a message to the screen.
watch_files_and_start_task(
  \() print(files have changed),
  "DESCRIPTION",
  \() dir("R", "*.R")
)
}
}
